LangSmith-Research-Agent/requirements.txt
langgraph
langchain-google-genai
langchain-core
langsmith
python-dotenv
requests
packaging<25
